Support imports without starting server—second attempt.
