+++ src/index.ts
@@ -1,9 +1,5 @@
-// Re-export the server and its types
-export { createServer } from "./mcp/index.js";
+// Re-export extractor types only
 export type { SimplifiedDesign } from "./extractors/types.js";
-export type { FigmaService } from "./services/figma.js";
-export { getServerConfig } from "./config.js";
-export { startServer } from "./cli.js";
 
 // Flexible extractor system
 export type {
