+++ tsup.config.ts
@@ -5,7 +5,7 @@ const packageVersion = process.env.npm_package_version;
 
 export default defineConfig({
   clean: true,
-  entry: ["src/index.ts", "src/cli.ts"],
+  entry: ["src/index.ts", "src/bin.ts", "src/mcp-server.ts"],
   format: ["esm"],
   minify: !isDev,
   target: "esnext",
@@ -13,7 +13,7 @@ export default defineConfig({
   outExtension: ({ format }) => ({
     js: ".js",
   }),
-  onSuccess: isDev ? "node dist/cli.js" : undefined,
+  onSuccess: isDev ? "node dist/bin.js" : undefined,
   define: {
     "process.env.NPM_PACKAGE_VERSION": JSON.stringify(packageVersion),
   },
